Split DungeonRouteControllerTest into per-endpoint files
Addresses review feedback that the single 1k-line region-based file was
too long. Splits into a DungeonRouteController/ directory (mirroring the
APICombatLogController test convention) with a shared
DungeonRouteControllerCreateTestBase holding the common dungeon-fetching
helpers, and one file per concern: SaveNewTemporary, SaveNew, Throttle.
